+++ lib/args/args_utils.dart
@@ -1,12 +1,26 @@
 abstract class ArgsUtils {
-  /// Commands.
-  static const watch = 'watch';
-  static const build = 'build';
-  static const prune = 'prune';
+  static const commands = _Commands();
+  static const args = _Args();
+}
 
-  /// Argument flags & options.
-  static const verbose = 'verbose';
-  static const debug = 'debug';
-  static const cacheDirectory = 'cache-directory';
-  static const lockPrune = 'prune';
+class _Commands {
+  String get watch => 'watch';
+  String get build => 'build';
+  String get cache => 'cache';
+
+  // Cache subcomammnds
+  String get prune => 'prune';
+  String get list => 'list';
+
+  const _Commands();
+}
+
+class _Args {
+  String get verbose => 'verbose';
+  String get debug => 'debug';
+  String get cacheDirectory => 'cache-directory';
+  String get lockPrune => 'prune';
+  String get clear => 'clear';
+
+  const _Args();
 }