+++ lib/args/build_and_watch_argument_parser.dart
@@ -3,10 +3,10 @@ import 'package:cached_build_runner/args/args_utils.dart';
 import 'package:cached_build_runner/utils/logger.dart';
 import 'package:cached_build_runner/utils/utils.dart';
 
-class ArgumentParser {
+class BuildAndWatchArgumentParser {
   final ArgParser _argParser;
 
-  ArgumentParser(this._argParser) {
+  BuildAndWatchArgumentParser(this._argParser) {
     _addFlagAndOption();
   }
 
@@ -17,46 +17,46 @@ class ArgumentParser {
     final result = _argParser.parse(arguments);
 
     /// cache directory
-    if (result.wasParsed(ArgsUtils.cacheDirectory)) {
-      Utils.appCacheDirectory = result[ArgsUtils.cacheDirectory] as String;
+    if (result.wasParsed(ArgsUtils.args.cacheDirectory)) {
+      Utils.appCacheDirectory = result[ArgsUtils.args.cacheDirectory] as String;
     } else {
       Utils.appCacheDirectory = Utils.getDefaultCacheDirectory();
       Logger.i(
-        "As no '${ArgsUtils.cacheDirectory}' was specified, using the default directory: ${Utils.appCacheDirectory}",
+        "As no '${ArgsUtils.args.cacheDirectory}' was specified, using the default directory: ${Utils.appCacheDirectory}",
       );
     }
 
     /// verbose
-    Utils.isVerbose = result[ArgsUtils.verbose] as bool;
-    Utils.isDebug = result[ArgsUtils.debug] as bool;
+    Utils.isVerbose = result[ArgsUtils.args.verbose] as bool;
+    Utils.isDebug = result[ArgsUtils.args.debug] as bool;
 
     // enable prunning
-    Utils.isPruneEnabled = result[ArgsUtils.lockPrune] as bool;
+    Utils.isPruneEnabled = result[ArgsUtils.args.lockPrune] as bool;
   }
 
   void _addFlagAndOption() {
     _argParser
       ..addFlag(
-        ArgsUtils.verbose,
+        ArgsUtils.args.verbose,
         abbr: 'v',
         help: 'Enables verbose mode',
         negatable: false,
       )
       ..addFlag(
-        ArgsUtils.debug,
+        ArgsUtils.args.debug,
         abbr: 'd',
         help: 'Enables debug mode',
         negatable: false,
       )
       ..addFlag(
-        ArgsUtils.lockPrune,
+        ArgsUtils.args.lockPrune,
         abbr: 'p',
         help: 'Enable pruning cache directory when pubspec.lock was changed since last build.',
         defaultsTo: true,
       )
       ..addSeparator('')
       ..addOption(
-        ArgsUtils.cacheDirectory,
+        ArgsUtils.args.cacheDirectory,
         abbr: 'c',
         help: 'Provide the directory where this tool can keep the caches.',
       );